Actually add group number to groups on a per-endpoint basis.

+++ src/mavlink-router/endpoint.h
@@ -111,7 +111,7 @@ public:
     void postprocess_msg(int target_sysid, int target_compid, uint8_t src_sysid, uint8_t src_compid, uint32_t msg_id);
 
     bool allowed_by_filter(uint32_t msg_id);
-    bool add_group(uint32_t group_id);
+    bool add_group(uint16_t group_id);
     void add_message_to_filter(uint32_t msg_id) { _message_filter.push_back(msg_id); }
     void add_message_to_nodelay(uint32_t msg_id) { _message_nodelay.push_back(msg_id); }
     bool allowed_by_dropout();
